Add the HuffmanCode class
This class offers a nice way to not have to track encoding and decoding
dictionaries for one set of messages.

The variable names and docstrings are still a little wonky. In my head,
strings aren't the only thing that could be encoded. In practice, I'm
not sure if it works.